+++ FrameScript/Components/WorkspaceLayout.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+enum WorkspaceLayout {
+    static func contentInset(isFocusModeEnabled: Bool) -> CGFloat {
+        isFocusModeEnabled ? 72 : 48
+    }
+}